scraper_backend/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.1.0
Automat==0.7.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.7
image==1.5.24
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
parsel==1.5.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.1.1
pylint-django==2.0
pylint-plugin-utils==0.4
pyOpenSSL==18.0.0
python-dotenv==0.9.1
python-scrapyd-api==2.1.2
pytz==2018.5
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
w3lib==1.19.0
whitenoise==4.0
wrapt==1.10.11
zope.interface==4.5.0
